refactor: checkNickname관련 소스들 features/user로 이동
